Simplify materialization refresh scheduling
Reuse the materialization queue version and consolidate refresh scheduling into one queued entry point.
