+++ src/data-providers/treecrdt/sync/index.ts
@@ -1,6 +1,3 @@
 /* eslint-disable import/prefer-default-export -- barrel re-exports */
 export { getTreecrdtSyncBaseUrl } from './config'
-export {
-  applyMaterializedThoughtsToStore,
-  enqueueMaterializedThoughtsToStore,
-} from './applyMaterializedThoughtsToStore'
+export { default as enqueueMaterializedThoughtsToStore } from './enqueueMaterializedThoughtsToStore'
